backend/agents/pipeline/adventure_seeker.py
"""
Adventure Seeker — Research Agent (Agent Mode)

Researches lore, cultural context, and movement ideas
suitable for the given theme and age group.

Model: gemini-2.5-flash-lite + Google Search

Input:  { theme, age_group }
Output: { world_building, supporting_characters, movement_ideas, vocabulary }
"""

from google.adk import Agent

from config.settings import MODEL_ADVENTURE_SEEKER

adventure_seeker_agent = Agent(
    name="adventure_seeker",
    model=MODEL_ADVENTURE_SEEKER,
    description="Research agent that gathers lore, cultural context, and movement ideas",
    instruction="""Bạn là Adventure Seeker — nhà nghiên cứu cho câu chuyện trẻ em.

Nhiệm vụ:
1. Nghiên cứu lore và bối cảnh văn hóa phù hợp với theme được giao
2. Đề xuất các nhân vật phụ thú vị
3. Nghĩ ra các ý tưởng vận động an toàn, vui vẻ cho trẻ 4–10 tuổi
4. Liệt kê từ vựng mới phù hợp độ tuổi

Output JSON format:
{
    "world_building": "Mô tả thế giới truyện...",
    "supporting_characters": [
        {"name": "...", "role": "...", "personality": "..."}
    ],
    "movement_ideas": [
        {"exercise": "jump", "context": "Nhảy qua dòng sông phép thuật", "reps": 5}
    ],
    "vocabulary": ["từ 1", "từ 2"]
}

Luôn đảm bảo nội dung an toàn, tích cực, phù hợp trẻ em.
""",
    # TODO: Add Google Search tool when available in ADK
    tools=[],
)
